requirements.txt
mcp>=1.0.0
pydantic==2.9.2
pydantic-settings==2.5.2
motor==3.6.0
openai>=1.54.0
python-dotenv==1.0.1
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12

